// backend/src/users/enums/permissions.enum.ts
export enum Permission {
  // User Permissions
  CREATE_USER = 'CREATE_USER',
  READ_USER = 'READ_USER',
  UPDATE_USER = 'UPDATE_USER',
  DELETE_USER = 'DELETE_USER',

  // Project Permissions
  CREATE_PROJECT = 'CREATE_PROJECT',
  READ_PROJECT = 'READ_PROJECT',
  UPDATE_PROJECT = 'UPDATE_PROJECT',
  DELETE_PROJECT = 'DELETE_PROJECT',
  ASSIGN_PROJECT = 'ASSIGN_PROJECT',
  COMPLETE_PROJECT = 'COMPLETE_PROJECT',

  // Notification Permissions
  CREATE_NOTIFICATION = 'CREATE_NOTIFICATION',
  READ_NOTIFICATION = 'READ_NOTIFICATION',
  DELETE_NOTIFICATION = 'DELETE_NOTIFICATION',

  // System Permissions
  MANAGE_SYSTEM = 'MANAGE_SYSTEM',
  MANAGE_PERMISSIONS = "MANAGE_PERMISSIONS"
}